Rename CSSSTests -> CSSS.Tests
