+++ CSSS.Tests/CSSS.Tests.csproj
@@ -1,0 +1,62 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectGuid>{4FEAF488-4BAB-47C0-A4BC-BA850512D34A}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <RootNamespace>CSSSTests</RootNamespace>
+    <AssemblyName>CSSSTests</AssemblyName>
+    <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
+    <ReleaseVersion>0.3</ReleaseVersion>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>bin\Debug</OutputPath>
+    <DefineConstants>DEBUG;</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release</OutputPath>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="System" />
+    <Reference Include="nunit.framework, Version=3.12.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
+      <HintPath>..\packages\NUnit.3.12.0\lib\net40\nunit.framework.dll</HintPath>
+    </Reference>
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="Test.cs" />
+    <Compile Include="BootstrapTests.cs" />
+    <Compile Include="InitTests.cs" />
+    <Compile Include="KernelTests.cs" />
+    <Compile Include="IssueChecks\IssueChecksTests.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\CSSS\CSSS.csproj">
+      <Project>{8346EB71-A8AE-41E0-BBE0-C93260F5AD31}</Project>
+      <Name>CSSS</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CSSSConfig\CSSSConfig.csproj">
+      <Project>{2B005BD2-677F-4646-BB17-673AD50DF07D}</Project>
+      <Name>CSSSConfig</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CSSSCheckerEngine\CSSSCheckerEngine.csproj">
+      <Project>{7DF38F4C-6B96-4E2F-A76C-8F21D976CE56}</Project>
+      <Name>CSSSCheckerEngine</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="IssueChecks\" />
+  </ItemGroup>
+  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
+</Project>
