+++ CSSS.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29020.237
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32929.385
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{429D7241-EDB8-4E8E-BFE5-E84B7D05A1D2}"
 	ProjectSection(SolutionItems) = preProject
@@ -17,7 +17,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CSSS", "CSSS\CSSS.csproj", "{8346EB71-A8AE-41E0-BBE0-C93260F5AD31}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CSSSTests", "CSSSTests\CSSSTests.csproj", "{4FEAF488-4BAB-47C0-A4BC-BA850512D34A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CSSS.Tests", "CSSS.Tests\CSSS.Tests.csproj", "{4FEAF488-4BAB-47C0-A4BC-BA850512D34A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CSSSSupportLibrary", "CSSSSupportLibrary\CSSSSupportLibrary.csproj", "{486E86DE-4D41-41B0-AC5F-C19E734464A3}"
 EndProject
