--- CSSSTests/packages.config
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NUnit" version="3.12.0" targetFramework="net40" />
-  <package id="NUnit.Console" version="3.10.0" targetFramework="net40" />
-  <package id="NUnit.ConsoleRunner" version="3.10.0" targetFramework="net40" />
-  <package id="NUnit.Extension.NUnitProjectLoader" version="3.6.0" targetFramework="net40" />
-  <package id="NUnit.Extension.NUnitV2Driver" version="3.7.0" targetFramework="net40" />
-  <package id="NUnit.Extension.NUnitV2ResultWriter" version="3.6.0" targetFramework="net40" />
-  <package id="NUnit.Extension.TeamCityEventListener" version="1.0.7" targetFramework="net40" />
-  <package id="NUnit.Extension.VSProjectLoader" version="3.8.0" targetFramework="net40" />
-</packages>
